pyproject.toml
[project]
name = "servicematrix"
version = "1.0.0"
description = "GitHubネイティブ × AI統治型 ITサービス統治基盤"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "slowapi>=0.1.9",
    "structlog>=24.2.0",
    "apscheduler>=3.10.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.40b0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "strawberry-graphql[fastapi]>=0.235.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "anyio[trio]>=4.3.0",
    "aiosqlite>=0.20.0",
]
postgres = [
    "asyncpg>=0.29.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ─── Poetry 互換（poetry install での開発時用） ───────────────────────────────

[tool.poetry]
name = "servicematrix"
version = "1.0.0"
description = "GitHubネイティブ × AI統治型 ITサービス統治基盤"
authors = ["ServiceMatrix Team"]
python = "^3.12"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
alembic = "^1.13.0"
asyncpg = "^0.29.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = ">=4.0.0,<5.0.0"
python-multipart = "^0.0.9"
httpx = "^0.27.0"
redis = "^5.0.0"
slowapi = "^0.1.9"
structlog = "^24.2.0"
apscheduler = "^3.10.0"
scikit-learn = ">=1.4.0"
numpy = ">=1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
ruff = "^0.4.0"
anyio = {extras = ["trio"], version = "^4.3.0"}
aiosqlite = "^0.20.0"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "S"]
ignore = ["S101", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml --cov-fail-under=75"
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*",
    "ignore::DeprecationWarning:fastapi.*",
]
markers = [
    "pg: PostgreSQL integration tests (require running postgres)",
]

[tool.coverage.run]
source = ["src"]
omit = ["src/models/*", "src/schemas/*", "src/__init__.py", "src/worker/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
no_strict_optional = true
exclude = ["alembic/", "tests/"]
